Add a SYSTEM_DOCTEST CMake option
This option is off by default, maintaining the previous behavior. When
enabled (along with FASTFLOAT_TEST), it bypasses the FetchContent
machinery for doctest so that a system-wide installation of the doctest
header can be easily used. In this case, the header doctest/doctest.h
should be available on the compiler’s include path.

This option is especially useful for Linux distributions and others that
need to run the tests in fully offline build environments.

Fixes #83.